+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [Unreleased] - ReleaseDate
+
+### Added
+
+* Ignoring bad blocks and other small IO issues.
+
 ## [v0.4.0] - 2020-06-10
 
 ### Added